feat: implement document loading and processing for DOCX and XLS formats
